fix(cli): clear screen on boot, show bootstrap spinner, fix orphaned tools (#57, #85, #86)
